Bug fix for 32 bit version. If compiled for 32 bit certain variables would not be sufficently large. This was a result of the assumption that size_t had the size 64 bit for all platforms but that was incorrect. As a result the comipler warnings would be given and the hashing function would give the incorrect result. The solution was to change certain usages of size_t with uint64_t.

+++ sha1.h
@@ -14,11 +14,11 @@
 typedef struct SHA1_CTX {
   uint32_t h[5];
   uint8_t block[BLOCK_BYTES];
-  size_t active_len;
-  size_t len;
+  uint64_t active_len;
+  uint64_t len;
 } SHA1_CTX;
 
 void SHA1_Init(SHA1_CTX *ctx);
-void SHA1_Update(SHA1_CTX *ctx, const void *data, size_t len);
+void SHA1_Update(SHA1_CTX *ctx, const void *data, uint64_t len);
 void SHA1_Final(SHA1_CTX *ctx, unsigned char *message_digest);
 #endif
